text-dream/python/linear_classifier/requirements.txt
torch
pytorch_pretrained_bert
absl_py
